Use unshare.HomeDir() when retrieving HOME environment
This change will also lookup the homedir in the /etc/passwd file
if $HOME does not exists.

This Patch also copies homedir_linux.go to homedir_unix.go, since these
functions can be used for Linux and Darwin systems.

homedir.Get() should also check for the rootless user not just the current
user, it will get fooled when running in a user namespace.

These functions work fine on the MAC and are used by Podman.
